+++ src/app/dashboard/layout.tsx
@@ -3,18 +3,14 @@ import DashboardNav from "@/components/dashboard/dashboardNav"
 
 export default function DashboardLayout({
 	children,
-
 }: {
 	children: React.ReactNode,
-
 }) {
 	return (
 		<section className="grid h-full grid-cols-[auto_minmax(400px,_1fr)]">
-
 			<DashboardNav />
 			{children}
 			<CreateEvent />
-
 		</section>
 	)
 }
